+++ HISTORY.rst
@@ -3,6 +3,11 @@ History
 =======
 
 
+0.X.X (XXXX-XX-XX)
+------------------
+*
+
+
 0.8.0 (2021-11-10)
 ------------------
 * AIQPD has been renamed "Quantile-Preserving, Localized Analogs Downscaling" (QPLAD). AIQPD-named commands have been switch to QPLAD. This is backward compatibility breaking. (PR #131, @brews)